+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Type coercion issue in `sum_results` by using dynamic default values based on inferred types
 
 ## [0.3.10] - 2026-01-11